+++ src/kind/delete.go
@@ -2,11 +2,11 @@ package kind
 
 import (
 	"fmt"
-	"infra-lab-cli/src/common"
+	"infra-lab-cli/src/utils"
 )
 
 func deleteCluster(binaryName, clusterName string) (err error) {
-	_, _, err = common.ExecBinaryCommand(
+	_, _, err = utils.ExecBinaryCommand(
 		binaryName,
 		fmt.Sprintf("delete cluster --name %s", clusterName),
 		true,
@@ -21,8 +21,8 @@ func deleteCluster(binaryName, clusterName string) (err error) {
 }
 
 func DeleteCluster(binaryName string, cluster Cluster) (err error) {
-	if !common.IsBinaryInPath(binaryName) {
-		fmt.Print(common.BinaryNotFoundError(binaryName))
+	if !utils.IsBinaryInPath(binaryName) {
+		fmt.Print(utils.BinaryNotFoundError(binaryName))
 		return nil
 	}
 
@@ -31,7 +31,7 @@ func DeleteCluster(binaryName string, cluster Cluster) (err error) {
 		return err
 	}
 
-	if common.IfStringInSlice(cluster.Name, clusters) {
+	if utils.IfStringInSlice(cluster.Name, clusters) {
 		err = deleteCluster(binaryName, cluster.Name)
 		if err != nil {
 			return err
