+++ src/minikube/start.go
@@ -2,11 +2,11 @@ package minikube
 
 import (
 	"fmt"
-	"infra-lab-cli/src/common"
+	"infra-lab-cli/src/utils"
 )
 
 func startCluster(binaryName string, cluster Cluster) (err error) {
-	_, _, err = common.ExecBinaryCommand(
+	_, _, err = utils.ExecBinaryCommand(
 		binaryName,
 		fmt.Sprintf("-p %s start", cluster.Name),
 		true,
@@ -21,8 +21,8 @@ func startCluster(binaryName string, cluster Cluster) (err error) {
 }
 
 func StartCluster(binaryName string, cluster Cluster) error {
-	if !common.IsBinaryInPath(binaryName) {
-		fmt.Print(common.BinaryNotFoundError(binaryName))
+	if !utils.IsBinaryInPath(binaryName) {
+		fmt.Print(utils.BinaryNotFoundError(binaryName))
 		return nil
 	}
