+++ src/podman/inspectMachine.go
@@ -3,12 +3,12 @@ package podman
 import (
 	"encoding/json"
 	"fmt"
-	"infra-lab-cli/src/common"
+	"infra-lab-cli/src/utils"
 	"strings"
 )
 
 func InspectMachine(binaryName string, machineName string) (machine *InspectedMachine, err error) {
-	stdout, _, err := common.ExecBinaryCommand(
+	stdout, _, err := utils.ExecBinaryCommand(
 		binaryName,
 		fmt.Sprintf("machine inspect %s", machineName),
 		false,
