+++ src/podman/start.go
@@ -2,16 +2,16 @@ package podman
 
 import (
 	"fmt"
-	"infra-lab-cli/src/common"
+	"infra-lab-cli/src/utils"
 )
 
 func StartMachine(binaryName, machineName string) (err error) {
-	if !common.IsBinaryInPath(binaryName) {
-		fmt.Print(common.BinaryNotFoundError(binaryName))
+	if !utils.IsBinaryInPath(binaryName) {
+		fmt.Print(utils.BinaryNotFoundError(binaryName))
 		return nil
 	}
 
-	_, _, err = common.ExecBinaryCommand(
+	_, _, err = utils.ExecBinaryCommand(
 		binaryName,
 		fmt.Sprintf("machine start %s", machineName),
 		true,
