+++ src/utils/utils_test.go
@@ -1,4 +1,4 @@
-package common
+package utils
 
 import (
 	"testing"
